backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.6.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
django-ratelimit==4.1.0
bcrypt==4.1.2
cryptography==41.0.8
